Update download recipients to handle additional and selected recipients
Lift the logic to handle additional and selected recipients into its own function to be shared by the recipients and download services.
