+++ app/services/notices/setup/download-recipients.service.js
@@ -10,6 +10,7 @@ const DownloadRecipientsPresenter = require('../../../presenters/notices/setup/d
 const FetchAbstractionAlertRecipientsService = require('./fetch-abstraction-alert-recipients.service.js')
 const FetchDownloadRecipientsService = require('./fetch-download-recipients.service.js')
 const SessionModel = require('../../../models/session.model.js')
+const RecipientsAddService = require('./recipients-add.service.js')
 
 /**
  * Orchestrates fetching and formatting the data needed for the notices setup download link
@@ -33,13 +34,11 @@ async function go(sessionId) {
 
     formattedData = AbstractionAlertDownloadRecipientsPresenter.go(recipients, session)
   } else {
-    const recipients = await FetchDownloadRecipientsService.go(session)
+    const recipientsData = await FetchDownloadRecipientsService.go(session)
 
-    const recipientsData = _additionalRecipients(recipients, session)
+    const recipients = RecipientsAddService.go(session, recipientsData, false)
 
-    const selectedRecipientsData = _selectedRecipients(session.selectedRecipients, recipientsData)
-
-    formattedData = DownloadRecipientsPresenter.go(selectedRecipientsData, session.notificationType)
+    formattedData = DownloadRecipientsPresenter.go(recipients, session.notificationType)
   }
 
   return {
@@ -47,20 +46,6 @@ async function go(sessionId) {
     type: 'text/csv',
     filename: `${notificationType} - ${referenceCode}.csv`
   }
-}
-
-function _additionalRecipients(recipientsData, session) {
-  if (session.additionalRecipients) {
-    return [...recipientsData, ...session.additionalRecipients]
-  }
-
-  return recipientsData
-}
-
-function _selectedRecipients(selectedRecipients, recipientsData) {
-  return recipientsData.filter((recipient) => {
-    return selectedRecipients.includes(recipient.contact_hash_id)
-  })
 }
 
 module.exports = {
